Ensure useFieldControlValidation is called once
